% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{find_edges_shortest_path}
\alias{find_edges_shortest_path}
\title{find shortest path}
\usage{
find_edges_shortest_path()
}
\description{
finds the nodes that make a shortest path between two given sets of nodes
and a distance paramter. This is done by breaking any edges between the sets of
nodes provided, and then asking for the shortest path from a node in one set to
the nodes in the other set, trimming to those that are within a specific number
of hops.
}
\details{
This function should use an igraph based graph, and then use some of the functionality
in all_shortest_paths
}
