% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{find_intersecting_nodes}
\alias{find_intersecting_nodes}
\title{alternative find_edges from both}
\usage{
find_intersecting_nodes(in_graph, start_nodes, end_nodes)
}
\arguments{
\item{in_graph}{a graphBAM}

\item{start_nodes}{the start nodes to use}

\item{end_nodes}{the end nodes to reach}
}
\value{
with new graph and all nodes in graph
}
\description{
a less general approach is to come from both the start and end nodes, go
one hop, and find the intersection of all pairwise comparisons. This method is
useful as it provides a simple check on the original find_edges method for 2
hops.
}
\examples{
library(STRINGDatabaseManipulation)
library(graph)
set.seed(1234)
link_data <- STRING10_links
link_data <- link_data[sample(nrow(link_data), 10000),]
in_graph <- string_2_graphBAM(link_data)
start_nodes <- end_nodes <- sample(nodes(link_graph), 100)
out_graph <- find_intersecting_nodes(in_graph, start_nodes, end_nodes)
}
