% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_processing.R
\name{process_string_aliases}
\alias{process_string_aliases}
\title{process STRING alias files}
\usage{
process_string_aliases(string_aliases)
}
\arguments{
\item{string_aliases}{the alias file}
}
\value{
data.frame
}
\description{
given a STRING alias file, generate a data.frame that can be used to map
various symbols
}
