% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_processing.R
\name{process_string_data}
\alias{process_string_data}
\title{process raw STRING data}
\usage{
process_string_data(string_file)
}
\arguments{
\item{string_file}{the data file}
}
\value{
data.frame
}
\description{
given a STRING data file, process it and save it for future use
}
