% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_processing.R
\name{process_string_links}
\alias{process_string_links}
\title{process raw STRING link data}
\usage{
process_string_links(string_links)
}
\arguments{
\item{string_links}{the links file}
}
\value{
data.frame
}
\description{
given a STRING links file, process it and save it for future use
}
