% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_files.R
\docType{data}
\name{STRING10_aliases}
\alias{STRING10_aliases}
\title{STRING V10 aliases for human}
\format{A data.frame with 2449433 rows and 3 variables:
\describe{
  \item{string}{the STRING ID}
  \item{other}{the other alias}
  \item{type}{the source of the alias (space separated for multiple sources)}}}
\source{
\url{http://string-db.org/newstring_download/protein.aliases.v10/9606.protein.aliases.v10.txt.gz}
Processed by RMF
}
\usage{
STRING10_aliases
}
\description{
Dataset containing various STRING aliases.
}
\keyword{datasets}
