% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_files.R
\docType{data}
\name{STRING10_links}
\alias{STRING10_links}
\title{STRING V10 detailed links for human}
\format{A data.frame with 8548002 rows and 10 variables:
\describe{
  \item{protein1}{first node in edge}
  \item{protein2}{second node in edge}
  \item{neighborhood}{evidence score for neighborhood}
  \item{fusion}{evidence score for fusion}
  \item{coocucurence}{evidence score for cooccurence}
  \item{coexpression}{evidence score from coexpression}
  \item{experimental}{score for experimental evidence}
  \item{database}{score for database evidence}
  \item{textmining}{score from textmining}
  \item{combined_score}{total combined score from all evidences}}}
\source{
\url{http://string-db.org/newstring_download/protein.links.detailed.v10/9606.protein.links.detailed.v10.txt.gz}
Processed by RMF
}
\usage{
STRING10_links
}
\description{
Dataset containing the detailed links from v10 of the STRING database for
human.
}
\keyword{datasets}
