% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_files.R
\docType{data}
\name{STRING11.5_9606_aliases}
\alias{STRING11.5_9606_aliases}
\title{STRING11.5_9606_aliases}
\format{
A data.frame with 58422 rows and 3 variables:
\describe{
\item{string}{the STRING ID}
\item{other}{the other alias}
\item{type}{the source of the alias}}

Processed by RMF on 2023-02-10
}
\source{
\url{https://stringdb-static.org/download/protein.aliases.v11.5/9606.protein.aliases.v11.5.txt.gz}
}
\usage{
STRING11.5_9606_aliases
}
\description{
Dataset containing Entrez and RecName aliases.
}
\keyword{datasets}
