% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_files.R
\docType{data}
\name{STRING11_9606_aliases}
\alias{STRING11_9606_aliases}
\title{STRING11_9606_aliases}
\format{A data.frame with 10000 rows and 3 variables:
\describe{
  \item{string}{the STRING ID}
  \item{other}{the other alias}
  \item{type}{the source of the alias (space separated for multiple sources)}}

Processed by RMF}
\source{
\url{https://stringdb-static.org/download/protein.aliases.v11.0/9606.protein.aliases.v11.0.txt.gz}
}
\usage{
STRING11_9606_aliases
}
\description{
Dataset containing various STRING aliases.
}
\keyword{datasets}
