% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_files.R
\docType{data}
\name{STRING11_9606_links}
\alias{STRING11_9606_links}
\title{STRING11_9606_links}
\format{A data.frame with 10000 rows and 10 variables:
\describe{
  \item{protein1}{first node in edge}
  \item{protein2}{second node in edge}
  \item{neighborhood}{evidence score for neighborhood}
  \item{fusion}{evidence score for fusion}
  \item{coocucurence}{evidence score for cooccurence}
  \item{coexpression}{evidence score from coexpression}
  \item{experimental}{score for experimental evidence}
  \item{database}{score for database evidence}
  \item{textmining}{score from textmining}
  \item{combined_score}{total combined score from all evidences}}

  Processed by RMF}
\source{
\url{https://stringdb-static.org/download/protein.links.detailed.v11.0/9606.protein.links.detailed.v11.0.txt.gz}
}
\usage{
STRING11_9606_links
}
\description{
Dataset containing the detailed links from v11 of the STRING database for
human. combined_score >= 400, and 10000 randomly chosen links. THIS IS NOT
THE COMPLETE DATASET!
}
\keyword{datasets}
