% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{string_2_graphBAM}
\alias{string_2_graphBAM}
\title{convert to graphBAM}
\usage{
string_2_graphBAM(string_data, use_weights = NULL)
}
\arguments{
\item{string_data}{a data.frame of STRING links}

\item{use_weights}{a column of the data.frame to use for weights. If NULL, value of 1 is used.}
}
\value{
graphBAM graph
}
\description{
convert a STRING data.frame to a graphBAM graph for use in other applications
}
