% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nodes_n_hops.R
\name{string_2_tidygraph}
\alias{string_2_tidygraph}
\title{convert to tidygraph}
\usage{
string_2_tidygraph(string_data, use_weights = NULL)
}
\arguments{
\item{string_data}{a data.frame of STRING links}

\item{use_weights}{a column of the data.frame to use for weights. If NULL, value of 1 is used.}
}
\value{
tidygraph
}
\description{
convert a STRING data.frame to a tidygraph for use in other applications
}
