% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{strip_species}
\alias{strip_species}
\title{strip species id from identifier}
\usage{
strip_species(string_id)
}
\arguments{
\item{string_id}{character vector of STRING IDs (XXXX.ENSPXXX)}
}
\value{
character vector
}
\description{
given a set of STRING ID's, remove the species taxonomy id part to get an ENSEMBL
protein ID.
}
