% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_processing.R
\name{process_string_aliases}
\alias{process_string_aliases}
\title{process STRING alias files}
\usage{
process_string_aliases(string_aliases)
}
\arguments{
\item{string_aliases}{the alias file}
}
\value{
data.frame
}
\description{
given a STRING alias file, generate a data.frame that can be used to map
various symbols.
}
\details{
The STRING alias file actually doesn't have a header, so the names we use
here for the columns are:
\itemize{
\item \strong{string}: the STRING ID, as ORG1.ENSPXXXXXX, \emph{9606.ENSP00000269305} is human, TP53
\item \strong{other}: the other ID, whatever that is, \emph{TP53} for this example
\item \strong{type}: the source of the alias, \emph{Ensembl_EntrezGene} in this case
}
}
