% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_processing.R
\name{process_string_id}
\alias{process_string_id}
\title{process STRING ID files}
\usage{
process_string_id(string_file)
}
\arguments{
\item{string_file}{the id file}
}
\value{
data.frame
}
\description{
given a STRING ID file, generate a data.frame that can be used to map
various symbols
}
