% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{string_2_graphBAM}
\alias{string_2_graphBAM}
\title{convert to graphBAM}
\usage{
string_2_graphBAM(string_data)
}
\arguments{
\item{string_data}{a data.frame of STRING links}
}
\value{
graphBAM graph
}
\description{
convert a STRING data.frame to a graphBAM graph for use in other applications
}
